Feature #加入跟踪一般直线 1. 加入可在上位机定义一般直线 2. 加入跟踪一般直线算法 3. 将除了经纬度，其他改为float变量

+++ HUST-1-Demo/Model/ShipData.cs
@@ -12,33 +12,33 @@ namespace HUST_1_Demo.Model
         public byte ShipID { get; set; }//船号
         public double Lat { get; set; }//纬度
         public double Lon { get; set; }//经度
-        public double pos_X { get; set; }//当前X坐标，此为测量坐标系下的坐标，北向为X，东向为Y
-        public double pos_Y { get; set; }//Y坐标
-        public double last_pos_X { get; set; }//上一坐标值
-        public double last_pos_Y { get; set; }//Y坐标
-        public double X_mm { get; set; }//图上X坐标
-        public double Y_mm { get; set; }//图上Y坐标
-        public double Control_Phi { get; set; }//当前用于控制采用的状态航向（航迹角或者航向角）
-        public double GPS_Phi { get; set; }//GPS-Phi航迹角
+        public float pos_X { get; set; }//当前X坐标，此为测量坐标系下的坐标，北向为X，东向为Y
+        public float pos_Y { get; set; }//Y坐标
+        public float last_pos_X { get; set; }//上一坐标值
+        public float last_pos_Y { get; set; }//Y坐标
+        public float X_mm { get; set; }//图上X坐标
+        public float Y_mm { get; set; }//图上Y坐标
+        public float Control_Phi { get; set; }//当前用于控制采用的状态航向（航迹角或者航向角）
+        public float GPS_Phi { get; set; }//GPS-Phi航迹角
         public float phi { get; set; }//航向角惯导
         public float Init_Phi { get; set; }//惯导原始数据，由于进行惯导0°初始化时需要获取当前惯导原始数据，所以这里需要保存惯导原始数据
         public float Phi_buchang { get; set; }//惯导补偿值
-        public double Pos_Phi { get; set; }//位置坐标角（极坐标，范围[-180, 180])
-        public double rud { get; set; }//舵角
+        public float Pos_Phi { get; set; }//位置坐标角（极坐标，范围[-180, 180])
+        public float rud { get; set; }//舵角
         public float speed { get; set; }//船速
         public string Time { get; set; }//时间
         public int gear { get; set; }//速度档
         public int CtrlRudOut { get; set; }//舵角控制输出量
         public int CtrlSpeedOut { get; set; }//速度控制输出量
-        public double Err_phi_In { get; set; }//积分控制中的积分量
+        public float Err_phi_In { get; set; }//积分控制中的积分量
 
         //控制参数
-        public double Kp { get; set; }
-        public double Ki { get; set; }
-        public double Kd { get; set; }
-        public double K1 { get; set; }
-        public double K2 { get; set; }
-        public double dl_err { get; set; }
+        public float Kp { get; set; }
+        public float Ki { get; set; }
+        public float Kd { get; set; }
+        public float K1 { get; set; }
+        public float K2 { get; set; }
+        public float dl_err { get; set; }
 
 
         public static double a = 6378137.0;//定义地球长半轴长度  
@@ -90,10 +90,10 @@ namespace HUST_1_Demo.Model
             Lat = Filter(FilterLat);
             Lon = Filter(FilterLon);*/
 
-            pos_X = (Lat - lat_start) * a * (1 - Math.Pow(earth_e, 2)) * 3.1415926 / (180 * Math.Sqrt(Math.Pow((1 - Math.Pow(earth_e * Math.Sin(Lat / 180 * 3.1415926), 2)), 3)));
-            pos_Y = -((Lon - lon_start) * a * Math.Cos(Lat / 180 * 3.1415926) * 3.1415926 / (180 * Math.Sqrt(1 - Math.Pow(earth_e * Math.Sin(Lat / 180 * 3.1415926), 2))));//Y坐标正向朝西
+            pos_X = (float)((Lat - lat_start) * a * (1 - Math.Pow(earth_e, 2)) * 3.1415926 / (180 * Math.Sqrt(Math.Pow((1 - Math.Pow(earth_e * Math.Sin(Lat / 180 * 3.1415926), 2)), 3))));
+            pos_Y = (float)(-((Lon - lon_start) * a * Math.Cos(Lat / 180 * 3.1415926) * 3.1415926 / (180 * Math.Sqrt(1 - Math.Pow(earth_e * Math.Sin(Lat / 180 * 3.1415926), 2)))));//Y坐标正向朝西
             
-            GPS_Phi = Math.Atan2(pos_Y - last_pos_Y, pos_X - last_pos_X) / Math.PI * 180;//航迹角
+            GPS_Phi = (float)(Math.Atan2(pos_Y - last_pos_Y, pos_X - last_pos_X) / Math.PI * 180);//航迹角
             last_pos_X = pos_X;//更新上一次坐标信息
             last_pos_Y = pos_Y;
